fix(sandbox): Harden egress proxy validation
Keep the public egress handler as a thin route adapter and move proxy/OIDC internals into sandbox-owned modules.

Use a trusted deployment audience for Vercel OIDC verification and refresh egress network policy when restoring sandboxes from an id hint.
